Use typenameExtractor to get types
