+++ packages/fabrix/src/index.ts
@@ -3,7 +3,7 @@ export { useFabrixContext } from "@context";
 export { gql } from "graphql-tag";
 export { FabrixComponent } from "@renderer";
 export { useFabrixClient, useDataFetch } from "@fetcher";
-export { type FieldType } from "@renderers/shared";
+export { type FieldType } from "@renderers/typename";
 export { type SubField } from "@renderers/fields";
 export {
   type FieldComponentEntry,
